Refactored discovery package
[#105] added main entry point `Discovery` struct which encapsualtes
peers and providers, added providers to abstract different discovery
techiques, refactored gRPC service and client to seed host and seed
provider.
